# virta-rest-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Virta Companies and Stations API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get a list of companies",
                "operationId": "54f966c056050f9db4fff6ee5a45a76a",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create a new company",
                "operationId": "00e2db7b0f71c32b377f4dfc4dd685a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Company Name"
                                    },
                                    "parent_company_id": {
                                        "type": "integer",
                                        "default": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "The validation error messages."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/companies/{company}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get a single company",
                "operationId": "dbcfdfb0281353ca1b00a67c0b0c2308",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "The ID of the company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update an existing company",
                "operationId": "17ee846405381b70d408b7532b44601a",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "The ID of the company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Company Name"
                                    },
                                    "parent_company_id": {
                                        "type": "integer",
                                        "default": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "The validation error messages."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete a company",
                "operationId": "815006ac9bb0fb97b1cad53b1a620652",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "The ID of the company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Company deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompanyResource": {},
...
        }
    }
}